Extract shared formatCurrency helper
PaymentHistoryRow previously rendered SGD/CAD/AUD/HKD/NZD/MXN as bare
`$`, indistinguishable from USD. Lift the dollar-family disambiguation
out of PremiumPrice into lib/formatCurrency.ts and use it from both
sites.
